+++ src/idl/mb.idl.php
@@ -497,7 +497,7 @@ DefineFunction(
     'name'   => "mb_ereg_search_getpos",
     'flags'  =>  HasDocComment,
     'return' => array(
-      'type'   => Int32,
+      'type'   => Int64,
       'desc'   => "mb_ereg_search_getpos() returns the point to start regular expression match for mb_ereg_search(), mb_ereg_search_pos(), mb_ereg_search_regs(). The position is represented by bytes from the head of string.",
     ),
   ));
