+++ src/idl/variable.idl.php
@@ -627,7 +627,7 @@ DefineFunction(
     'desc'   => "Import variables from an array into the current symbol table.\n\nChecks each key to see whether it has a valid variable name. It also checks for collisions with existing variables in the symbol table.",
     'flags'  =>  HasDocComment,
     'return' => array(
-      'type'   => Int32,
+      'type'   => Int64,
       'desc'   => "Returns the number of variables successfully imported into the symbol table.",
     ),
     'args'   => array(
